+++ crud-vendas-demo/src/main/java/com/leo/crud/vendas/projections/SaleReportProjectionRecord.java
@@ -1,4 +1,4 @@
 package com.leo.crud.vendas.projections;
 
-public record SaleReportProjectionRecord(String product, Integer amount, Double unitValue, Double total) {
+public record SaleReportProjectionRecord(String product, Long amount, Double unitValue, Double total) {
 }
